Cleanup stats generation
