+++ scripts/stats/package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
+    "stats": "node index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Chris Wiegman <chris.wiegman@wpengine.com>",
